fix future forecast values table bug v4

+++ Toronto_2024_Electricity_Demand_Forecast.py
@@ -154,6 +154,30 @@ def show_Toronto_2024_Electricity_Demand_Forecast():
 
 #     # ... [the previous part of your Streamlit app code] ...
 
+#     # Convert the 'ds' column to datetime if it's not already
+#     test_data_predictions['ds'] = pd.to_datetime(test_data_predictions['ds'])
+
+#     # Filter the DataFrame for dates that are today or in the future
+#     today = pd.Timestamp(datetime.now().date())  # Get today's date
+#     future_data = test_data_predictions[test_data_predictions['ds'] >= today]
+
+#     # Create the table display
+#     table_to_display = future_data.copy()
+#     table_to_display['Date'] = table_to_display['ds'].dt.date
+#     table_to_display['Time'] = table_to_display['ds'].dt.time
+
+#     # Round the 'XGBoost Forecast Demand' to two decimal places
+#     table_to_display['XGBoost Forecast Demand'] = table_to_display['y_predictions_xgb'].round(2)
+
+#     # Select and rename the columns for the table
+#     table_to_display = table_to_display[['Date', 'Time', 'XGBoost Forecast Demand']]
+
+#     # Use Streamlit to display the table without the index
+#     st.table(table_to_display.assign(hack='').set_index('hack'))
+    
+    
+    # ... [the previous part of your Streamlit app code] ...
+
     # Convert the 'ds' column to datetime if it's not already
     test_data_predictions['ds'] = pd.to_datetime(test_data_predictions['ds'])
 
@@ -166,11 +190,12 @@ def show_Toronto_2024_Electricity_Demand_Forecast():
     table_to_display['Date'] = table_to_display['ds'].dt.date
     table_to_display['Time'] = table_to_display['ds'].dt.time
 
-    # Round the 'XGBoost Forecast Demand' to two decimal places
-    table_to_display['XGBoost Forecast Demand'] = table_to_display['y_predictions_xgb'].round(2)
+    # Ensure 'y_predictions_xgb' is of float type, then round to two decimal places
+    table_to_display['XGBoost Forecast Demand'] = table_to_display['y_predictions_xgb'].astype(float).round(2)
 
     # Select and rename the columns for the table
     table_to_display = table_to_display[['Date', 'Time', 'XGBoost Forecast Demand']]
 
-    # Use Streamlit to display the table without the index
+    # Use Streamlit to display the table without the original index and the temporary 'hack' column
     st.table(table_to_display.assign(hack='').set_index('hack'))
+
